requirements.txt
pandas~=2.0.3
SQLAlchemy~=2.0.20
numpy~=1.25.2
python-dotenv~=1.0.0